+++ src/components/ui/sonner.tsx
@@ -1,5 +1,5 @@
 // Copyright (c) 2026. Licensed under AGPLv3.
-import { useTheme } from "next-themes";
+import { useTheme } from "@/context/theme-provider";
 import { Toaster as Sonner } from "sonner";
 
 type ToasterProps = React.ComponentProps<typeof Sonner>;